refactor : 메인페이지 로딩 컨트롤러 클래스 위치 변경, 인기글 로직 수정
